moves old class to old node where it belongs
